requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data processing
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.1.0
PyMuPDF==1.23.8
pydub==0.25.1

# AI & ML
openai==1.3.7
google-generativeai==0.3.2
tiktoken==0.5.1
sentence-transformers==2.2.2

# Database
supabase==2.0.0
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.5.0 